% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddate.R
\name{ddate}
\alias{ddate}
\title{Converting Greg's to POEE date.}
\usage{
ddate()
}
\arguments{
\item{""}{This function is serious business, it doesn't take arguments from anyone.}
}
\value{
The current day, season and YOLD. Except it's St. Tib's Day; it's St. Tib's Day then.
}
\description{
Converting Greg's to POEE date.
}

