% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddate_convert.R
\name{ddate_convert}
\alias{ddate_convert}
\title{Converting Greg's to POEE date.}
\usage{
ddate_convert(greg = Sys.Date())
}
\value{
The current day, season and YOLD. Except it's St. Tib's Day; it's St. Tib's Day then.
}
\description{
Converting Greg's to POEE date.
}

