% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnord.R
\name{fnord}
\alias{fnord}
\title{Fnord.}
\usage{
fnord()
}
\arguments{
\item{none}{really}
}
\value{
Sacred sermons of the POEE as scribed by pope Pixie Marauder I
}
\description{
Join the KSC now. Defnordify the world. Peace out.
}
\examples{
fnord()
}

