% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddate.R
\name{poee}
\alias{poee}
\title{Converting Greg's to POEE date.}
\usage{
poee(greg = Sys.Date())
}
\arguments{
\item{greg}{The date in standard ISO 8601 date format, e.g. \code{2016-02-18}.}
}
\value{
The current (or given) day, season and YOLD. Except it's St. Tib's Day; it's St. Tib's Day then.
}
\description{
Converting Greg's to POEE date.
}

