% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ye_olden_ddate.R
\name{ddate}
\alias{ddate}
\title{Printing the POEE date.}
\usage{
ddate()
}
\value{
The current day, season and YOLD. Except it's St. Tib's Day; it's St. Tib's Day then.
}
\description{
Printing the POEE date.
}

